+++ buzzvil-sdk/app/src/main/java/com/buzzvil/sample/buzzvilsdksample/Constant.kt
@@ -1,5 +1,10 @@
 package com.buzzvil.sample.buzzvilsdksample
 
+/**
+ * 여기에 버즈빌 매니저로부터 발급받은 ID를 입력해주세요.
+ * 유효하지 않은 ID를 입력하면 광고가 노출되지 않습니다.
+ * 올바른 ID를 입력해도 테스트 광고가 노출되지 않는 경우 버즈빌 매니저에게 문의해주세요.
+ */
 object Constant {
     const val YOUR_APP_ID = "YOUR_APP_ID"
     const val YOUR_FEED_ID = "YOUR_FEED_ID"
