+++ buzzvil-sdk/app/src/main/java/com/buzzvil/sample/buzzvilsdksample/benefithub/BenefitHubFragmentActivity.kt
@@ -1,12 +1,14 @@
-package com.buzzvil.sample.buzzvilsdksample
+package com.buzzvil.sample.buzzvilsdksample.benefithub
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.buzzvil.buzzad.benefit.feed.benefithub.BenefitHubFragment
+import com.buzzvil.sample.buzzvilsdksample.R
 import com.buzzvil.sample.buzzvilsdksample.databinding.ActivityBenefitHubFragmentBinding
 
 class BenefitHubFragmentActivity : AppCompatActivity() {
     private lateinit var binding: ActivityBenefitHubFragmentBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityBenefitHubFragmentBinding.inflate(layoutInflater)
