+++ buzzvil-sdk/app/src/main/res/layout/activity_benefit_hub_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BenefitHubFragmentActivity">
+    tools:context=".benefithub.BenefitHubFragmentActivity">
 
     <!-- 내비게이션 바를 원하는 형태로 구현합니다. -->
     <LinearLayout
