+++ buzzvil-sdk/app/src/main/res/layout/activity_main.xml
@@ -16,6 +16,15 @@
             android:orientation="vertical">
 
             <TextView
+                android:id="@+id/sdkVersionTextView"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="16dp"
+                android:layout_marginTop="16dp"
+                android:textSize="16dp"
+                tools:text="SDK Version: 5.0.0" />
+
+            <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="16dp"
@@ -123,84 +132,26 @@
                 android:text="Native"
                 android:textSize="24dp" />
 
-            <LinearLayout
-                android:layout_width="match_parent"
+            <Button
+                android:id="@+id/nativeSimpleButton"
+                android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="16dp"
-                android:orientation="horizontal">
-
-                <Button
-                    android:id="@+id/loadSimpleNativeAdButton"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Load Simple" />
-
-                <Button
-                    android:id="@+id/loadCardViewNativeAdButton"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="4dp"
-                    android:text="Load CardView" />
-            </LinearLayout>
-
-            <include
-                android:id="@+id/simpleNativeAdLayout"
-                layout="@layout/native_ad_simple" />
-
-            <include
-                android:id="@+id/cardViewNativeAdLayout"
-                layout="@layout/native_ad_card_view" />
+                android:text="Native Simple" />
 
             <Button
-                android:id="@+id/loadCarouselButton"
+                android:id="@+id/nativeCustomButton"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="16dp"
-                android:text="Load Carousel" />
+                android:text="Native Custom" />
 
-            <!-- 광고를 표시할 Carousel -->
-            <!-- clipToPadding, paddingHorizontal: 이전과 이후 아이템을 살짝 보여주기 위해 설정 -->
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/carouselRecyclerView"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:clipToPadding="false"
-                android:orientation="horizontal"
-                android:paddingHorizontal="32dp"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                tools:listitem="@layout/carousel_item" />
-
-            <!-- Carousel 하단에서 Feed를 열 수 있게 하는 Carousel To Feed Link -->
-            <!-- buzzvilFeedEntryViewName의 값은 "carousel_to_feed_link"를 사용해야 합니다. -->
-            <com.buzzvil.buzzad.benefit.presentation.feed.entrypoint.FeedEntryView
-                android:id="@+id/toFeedLink"
+            <Button
+                android:id="@+id/nativeCarouselButton"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_gravity="right"
-                android:layout_margin="16dp"
-                app:buzzvilFeedEntryViewName="carousel_to_feed_link">
-
-                <!-- 내부 레이아웃은 자유롭게 설정할 수 있습니다. 아래는 LinearLayout으로 구현한 예시입니다. -->
-                <LinearLayout
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:orientation="horizontal">
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="포인트 더 받으러 가기"
-                        android:textColor="#3B82F6"
-                        android:textSize="14dp"
-                        android:textStyle="bold" />
-
-                    <ImageView
-                        android:layout_width="24dp"
-                        android:layout_height="24dp"
-                        android:src="@drawable/baseline_chevron_right_24"
-                        app:tint="#3B82F6" />
-                </LinearLayout>
-            </com.buzzvil.buzzad.benefit.presentation.feed.entrypoint.FeedEntryView>
+                android:layout_marginHorizontal="16dp"
+                android:text="Native Carousel" />
         </LinearLayout>
     </ScrollView>
 </FrameLayout>
