+++ buzzvil-sdk/app/src/main/res/layout/activity_native_custom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".nativead.NativeCustomActivity">
+
+    <Button
+        android:id="@+id/loadCardViewNativeAdButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_margin="16dp"
+        android:text="Load CardView" />
+
+    <include
+        android:id="@+id/cardViewNativeAdLayout"
+        layout="@layout/native_ad_card_view" />
+</LinearLayout>
