+++ buzzvil-sdk/app/src/main/res/layout/activity_native_simple.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".nativead.NativeSimpleActivity">
+
+    <Button
+        android:id="@+id/loadSimpleNativeAdButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_margin="16dp"
+        android:text="Load Simple" />
+
+    <include
+        android:id="@+id/simpleNativeAdLayout"
+        layout="@layout/native_ad_simple" />
+</LinearLayout>
