+++ buzzvil-sdk/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BuzzvilSdkSample</string>
+    <string name="app_name">Buzzvil SDK</string>
 </resources>
